#Refactor Make delete use delete request Removed /userId/evidenceId endpoints Make notifications consistent with other notifications

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/EvidenceController.java
@@ -1,6 +1,5 @@
 package nz.ac.canterbury.seng302.portfolio.controller;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import nz.ac.canterbury.seng302.portfolio.model.Project;
 import nz.ac.canterbury.seng302.portfolio.model.Evidence;
 import nz.ac.canterbury.seng302.portfolio.model.User;
@@ -18,6 +17,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.event.EventListener;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
@@ -27,12 +28,11 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.web.socket.messaging.SessionDisconnectEvent;
 
 import java.util.*;
 
-import static org.springframework.http.HttpStatus. *;
+import static org.springframework.http.HttpStatus.*;
 
 /**
  * Controller for the evidence page
@@ -57,7 +57,7 @@ public class EvidenceController {
 
     private final Logger logger = LoggerFactory.getLogger(EvidenceController.class);
 
-    private final HashMap<Integer, EvidenceNotification> editing = new HashMap<>();
+    private static Set<EvidenceNotification> editing = new HashSet<>();
 
     private static final String EVIDENCE = "evidence";
 
@@ -73,16 +73,12 @@ public class EvidenceController {
 
     private static final String IS_CURRENT_USER = "isCurrentUserEvidence";
 
-    public EvidenceController(EvidenceService evidenceService) {
-        this.evidenceService = evidenceService;
-    }
-
     /**
      * Updates the model with the correct list of evidence
      * @param userId UserId containing the desired evidence
      * @return Page fragment containing events
      */
-    @GetMapping(path="/evidence/{userId}")
+    @GetMapping(path="user/{userId}/evidence")
     public String getEvidencePage(
             @AuthenticationPrincipal AuthState principal,
             @PathVariable("userId") int userId,
@@ -109,22 +105,18 @@ public class EvidenceController {
 
     /**
      * Updates the model with the correct list of evidence and a selected evidence
-     * @param userId User ID containing the evidence
      * @param evidenceId ID of the selected evidence object
      * @return The selected evidence fragment
      */
-    @GetMapping(path="/evidence/{userId}/{evidenceId}")
+    @GetMapping(path="/evidence/{evidenceId}")
     public ModelAndView selectedEvidence(
-            @PathVariable int userId,
             @PathVariable int evidenceId,
             @AuthenticationPrincipal AuthState principal) {
         ModelAndView mv = new ModelAndView("evidenceFragments::selectedEvidence");
-        List<Evidence> listEvidence = evidenceService.getEvidenceByUserId(userId);
-        mv.addObject(LIST_EVIDENCE, listEvidence);
         try {
-            Evidence selectedEvidence = evidenceService.getEvidence(evidenceId);
-            mv.addObject("selectedEvidence", selectedEvidence);
-            mv.addObject(IS_CURRENT_USER, PrincipalUtils.getUserId(principal)==userId);
+            Evidence evidence = evidenceService.getEvidence(evidenceId);
+            mv.addObject("selectedEvidence", evidence);
+            mv.addObject(IS_CURRENT_USER, PrincipalUtils.getUserId(principal) == evidence.getOwnerId());
             mv.addObject(NOTIFICATIONS, editing);
         } catch (IncorrectDetailsException e) {
             mv = new ModelAndView("evidence::serverMessages", NOT_FOUND);
@@ -138,24 +130,20 @@ public class EvidenceController {
      * @param ra Redirect Attribute frontend message object
      * @return link of html page to display
      */
-    @PostMapping(path="/evidence/{userId}/saveEvidence")
+    @PostMapping(path="/evidence/")
     public String saveEvidence(
             @ModelAttribute EvidenceDTO evidenceDTO,
-            @PathVariable("userId") int userId,
             RedirectAttributes ra,
             @AuthenticationPrincipal AuthState principal,
             @Header("simpSessionId") String sessionId) {
         Evidence evidence = new Evidence(evidenceDTO);
         try {
-            if (PrincipalUtils.getUserId(principal) != userId) {
+            if (PrincipalUtils.getUserId(principal) != evidence.getOwnerId()) {
                 throw new IncorrectDetailsException("You may only create evidence on your own evidence page");
             }
-            evidence.setOwnerId(userId);
             evidenceService.verifyEvidence(evidence);
             String message = evidenceService.saveEvidence(evidence);
-            EvidenceNotification notification = new EvidenceNotification(evidence.getEvidenceId(), "saved",
-                    0, PrincipalUtils.getUserName(principal), userId, sessionId);
-            notifyEvidence(notification);
+            notifyEvidence(evidence.getOwnerId(), evidence.getEvidenceId(), "edited");
             ra.addFlashAttribute(MESSAGE_SUCCESS, message);
         } catch(IncorrectDetailsException e) {
             ra.addFlashAttribute(MESSAGE_DANGER, e.getMessage());
@@ -168,10 +156,9 @@ public class EvidenceController {
      * @param evidenceId ID of the selected evidence
      * @return evidence form fragment
      */
-    @GetMapping(path="/evidence/{userId}/{evidenceId}/editEvidence")
+    @GetMapping(path="/evidence/{evidenceId}/editEvidence")
     public ModelAndView editEvidence(
-            @PathVariable int evidenceId,
-            @PathVariable int userId) {
+            @PathVariable int evidenceId) {
         ModelAndView mv = new ModelAndView("evidence::evidenceForm");
         try {
             Evidence evidence = evidenceService.getEvidence(evidenceId);
@@ -194,12 +181,12 @@ public class EvidenceController {
      * @param userId of user to create a new Evidence
      * @return evidence form fragment
      */
-    @GetMapping(path="/evidence/{userId}/getNewEvidence")
-    public ModelAndView createEvidence(
-            @PathVariable int userId) {
+    @GetMapping(path="/evidence/getNewEvidence")
+    public ModelAndView createEvidence(@AuthenticationPrincipal AuthState principal) {
         ModelAndView mv = new ModelAndView("evidence::evidenceForm");
-        Evidence evidence = evidenceService.getNewEvidence(userId);
+        Evidence evidence = evidenceService.getNewEvidence(PrincipalUtils.getUserId(principal));
         List<Project> listProjects = projectService.getAllProjects();
+        
         mv.addObject(EVIDENCE, evidence);
         mv.addObject(LIST_PROJECTS, listProjects);
         mv.addObject("submissionImg", apiPrefix+"/icons/create-icon.svg");
@@ -213,25 +200,26 @@ public class EvidenceController {
      * @param ra redirect attribute of for displaying the message
      * @return the link to the HTML page
      */
-    @PostMapping(path="/evidence/{userId}/{evidenceId}/deleteEvidence")
-    public String deleteEvidence(
+    @DeleteMapping(path="/evidence")
+    public ResponseEntity<String> deleteEvidence(
             @PathVariable int evidenceId,
-            @PathVariable int userId,
             @AuthenticationPrincipal AuthState principal,
-            @Header("simpSessionId") String sessionId,
-            RedirectAttributes ra) {
+            @Header("simpSessionId") String sessionId) {
         try {
-            String message = evidenceService.deleteEvidence(evidenceId);
-            ra.addFlashAttribute(MESSAGE_SUCCESS, message);
-            EvidenceNotification notification = new EvidenceNotification(evidenceId, "deleted",
-                    0, PrincipalUtils.getUserName(principal), userId, sessionId);
+            Evidence evidence = evidenceService.getEvidence(evidenceId);
+    
+            if (!(PrincipalUtils.checkUserIsTeacherOrAdmin(principal) || evidence.getOwnerId() == PrincipalUtils.getUserId(principal))) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Insufficient Permissions");
+            }   
 
-            notifyEvidence(notification);
+            String message = evidenceService.deleteEvidence(evidenceId);
+
+            notifyEvidence(evidence.getOwnerId(), evidenceId, "deleted");
+            return ResponseEntity.status(HttpStatus.OK).body(message);
         } catch (IncorrectDetailsException e) {
-            ra.addFlashAttribute(MESSAGE_DANGER, e.getMessage());
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
         }
 
-        return "redirect:/evidence/{userId}";
     }
 
     /**
@@ -239,7 +227,7 @@ public class EvidenceController {
      * @param userId the ID of user whose evidence list is requested
      * @return evidenceList fragment
      */
-    @GetMapping(path="/evidence/{userId}/getEvidenceList")
+    @GetMapping(path="user/{userId}/evidence/getEvidenceList")
     public ModelAndView getEvidenceList(
             @PathVariable int userId,
             @AuthenticationPrincipal AuthState principal) {
@@ -260,46 +248,36 @@ public class EvidenceController {
      */
     @MessageMapping("/evidence/edit")
     public void editing(EvidenceNotification notification, @AuthenticationPrincipal WebSocketPrincipal principal, @Header("simpSessionId") String sessionId) {
-        notification.setUserUpdating(principal.getName());
+        notification.setUsername(principal.getName());
         notification.setSessionId(sessionId);
-        notifyEvidence(notification);
-        if (notification.getAction().equals("editing"))
-            editing.put(notification.getEvidenceId(), notification);
-        else
-            editing.remove(notification.getEvidenceId());
-
+        if (notification.isActive()) {
+            notifyEvidence(notification.getUserId(), notification.getEvidenceId(), "editing");
+            editing.add(notification);
+        } else {
+            notifyEvidence(notification.getUserId(), notification.getEvidenceId(), "finished");
+            editing.remove(notification);
+        }
     }
 
     /**
      * Sends an update message to all clients connected to the websocket
-     * @param notification encapsulated data to be exchanged between front and back end
+     * @param projectId
+     * @param eventId
+     * @param action
      */
-    private void notifyEvidence(EvidenceNotification notification) {
-        List<Evidence> listEvidence = evidenceService.getEvidenceByUserId(notification.getUserId());
-        int firstEvidenceId = listEvidence.isEmpty() ? 0 : listEvidence.get(0).getEvidenceId();
-        notification.setFirstEvidenceId(firstEvidenceId);
-        try {
-            template.convertAndSend("/element/evidence/" + notification.getUserId(), new ObjectMapper().writeValueAsString(notification));
-        } catch (JsonProcessingException e) {
-            logger.info("Unable to convert and send notification {}", notification);
-        }
+    private void notifyEvidence(int userId, int evidenceId, String action) {
+        template.convertAndSend("/element/evidence/" + userId, ("evidence" + evidenceId + " " + action));
     }
 
     /**
-     * Detects when a websocket disconnects to remove them from the list of editors
+     * Detects when a websocket disconnects to send remove them from the list of editors
      * @param event Websocket disconnect event
      */
     @EventListener
     public void onApplicationEvent(SessionDisconnectEvent event) {
-        for (Map.Entry<Integer, EvidenceNotification> current : editing.entrySet()) {
-            EvidenceNotification currentNotification = current.getValue();
-            if (currentNotification.getSessionId().equals(event.getSessionId())) {
-                currentNotification.setAction("finished");
-                notifyEvidence(currentNotification);
-                editing.remove(current.getKey());
-            }
-        }
-
+        editing.stream().filter(notification->notification.getSessionId().equals(event.getSessionId()))
+        .forEach(notification->template.convertAndSend("/element/evidence/" + notification.getEvidenceId(), ("evidence" + notification.getEvidenceId() + " finished")));
+        editing.removeIf(notification->notification.getSessionId().equals(event.getSessionId()));
     }
